requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
arrow==1.4.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bibtexparser==1.4.3
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.6.7
datasets==4.3.0
deprecated==1.2.18
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring-parser==0.17.0
docutils==0.21.2
fake-useragent==2.2.0
filelock==3.20.0
free-proxy==1.1.3
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
imagesize==1.4.1
iniconfig==2.3.0
instructor==1.12.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2


networkx==3.5
nltk==3.9.2
nodeenv==1.9.1
numpy==2.3.4
openai==1.109.1
orjson==3.11.4
ormsgpack==1.11.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
pre-commit==4.3.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==22.0.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
pymupdf==1.26.5
pymupdf4llm==0.0.27
pyparsing==3.2.5
pysocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
qdrant-client==1.15.1
ragas==0.3.7
rank-bm25==0.2.2
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
roman-numerals-py==3.1.0
scholarly==1.7.11
scikit-network==0.33.3
scipy==1.16.2
selenium==4.38.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.3
websocket-client==1.9.0
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
